Nuke `GenericType.TypeAbstraction`
