+++ analyzer/src/main/kotlin/com/github/cubuspl42/sigmaLang/analyzer/semantics/expressions/GenericConstructor.kt
@@ -1,22 +1,15 @@
 package com.github.cubuspl42.sigmaLang.analyzer.semantics.expressions
 
 import com.github.cubuspl42.sigmaLang.analyzer.evaluation.scope.DynamicScope
-import com.github.cubuspl42.sigmaLang.analyzer.evaluation.values.DictValue
-import com.github.cubuspl42.sigmaLang.analyzer.evaluation.values.TableValue
 import com.github.cubuspl42.sigmaLang.analyzer.evaluation.values.Thunk
 import com.github.cubuspl42.sigmaLang.analyzer.evaluation.values.Value
-import com.github.cubuspl42.sigmaLang.analyzer.evaluation.values.asType
 import com.github.cubuspl42.sigmaLang.analyzer.lazier
 import com.github.cubuspl42.sigmaLang.analyzer.semantics.StaticScope
 import com.github.cubuspl42.sigmaLang.analyzer.semantics.chainWithIfNotNull
 import com.github.cubuspl42.sigmaLang.analyzer.semantics.expressions.AbstractionConstructor.ArgumentDeclaration
-import com.github.cubuspl42.sigmaLang.analyzer.semantics.introductions.Declaration
 import com.github.cubuspl42.sigmaLang.analyzer.semantics.types.GenericType
 import com.github.cubuspl42.sigmaLang.analyzer.semantics.types.TupleType
 import com.github.cubuspl42.sigmaLang.analyzer.semantics.types.Type
-import com.github.cubuspl42.sigmaLang.analyzer.semantics.types.TypeAlike
-import com.github.cubuspl42.sigmaLang.analyzer.semantics.types.TypeType
-import com.github.cubuspl42.sigmaLang.analyzer.semantics.types.TypeVariable
 import com.github.cubuspl42.sigmaLang.analyzer.syntax.expressions.GenericConstructorTerm
 
 class GenericConstructor(
@@ -79,23 +72,8 @@ class GenericConstructor(
         DiagnosedAnalysis(
             analysis = Analysis(
                 inferredType = GenericType(
-                    parameterType = metaArgumentType,
-                    typeAbstraction = object : GenericType.TypeAbstraction {
-                        override fun apply(parameterTable: TableValue): Type {
-                            val typeVariableReplacer = buildTypeVariableReplacer(
-                                traitType = metaArgumentType,
-                                path = TypeVariable.Path.Root,
-                                specificationTable = parameterTable,
-                                traitDeclaration = metaArgumentDeclaration,
-                            )
-
-                            val specifiedType = inferredBodyType.replaceType(
-                                typeReplacer = typeVariableReplacer,
-                            ) as Type
-
-                            return specifiedType
-                        }
-                    }
+                    parameterDeclaration = metaArgumentDeclaration,
+                    bodyType = inferredBodyType,
                 )
             ),
             directErrors = emptySet(), // TODO
@@ -110,56 +88,3 @@ class GenericConstructor(
     ): Thunk<Value> = body.bindDirectly(dynamicScope = dynamicScope)
 }
 
-private fun buildTypeVariableReplacer(
-    traitDeclaration: Declaration,
-    path: TypeVariable.Path,
-    traitType: TupleType,
-    specificationTable: TableValue,
-): TypeAlike.TypeReplacer = TypeAlike.TypeReplacer.combineAll(
-    replacers = traitType.entries.map { entry ->
-        val entryKey = entry.key
-        val specificationValue = specificationTable.read(entryKey)!!.value!!
-
-        buildTypeVariableReplacer(
-            traitDeclaration = traitDeclaration,
-            entryPath = path.extend(entryKey),
-            traitEntryType = entry.type,
-            specificationValue = specificationValue,
-        )
-    },
-)
-
-private fun buildTypeVariableReplacer(
-    traitDeclaration: Declaration,
-    entryPath: TypeVariable.Path,
-    traitEntryType: TypeAlike,
-    specificationValue: Value,
-): TypeAlike.TypeReplacer = when (traitEntryType) {
-    TypeType -> {
-        val specificationType = specificationValue.asType!!
-
-        object : TypeAlike.TypeReplacer {
-            override fun replace(
-                type: TypeAlike,
-            ): TypeAlike? =
-                if (type is TypeVariable && type.traitDeclaration == traitDeclaration && type.path == entryPath) {
-                    specificationType
-                } else {
-                    null
-                }
-        }
-    }
-
-    is TupleType -> {
-        val innerSpecificationTable = specificationValue as TableValue
-
-        buildTypeVariableReplacer(
-            traitDeclaration = traitDeclaration,
-            path = entryPath,
-            traitType = traitEntryType,
-            specificationTable = innerSpecificationTable,
-        )
-    }
-
-    else -> throw UnsupportedOperationException("Invalid trait")
-}
